+++ Controller/Payment/Callback/Dmn.php
@@ -144,6 +144,11 @@ class Dmn extends \Magento\Framework\App\Action\Action implements \Magento\Frame
             
             $this->moduleConfig->createLog($params, 'DMN params:');
             
+            if (!empty($params['type']) && 'CARD_TOKENIZATION' == $params['type']) {
+                $this->jsonOutput->setData('DMN report - this is Card Tokenization DMN.');
+                return $this->jsonOutput;
+            }
+            
             ### DEBUG
 //            $this->jsonOutput->setData('DMN manually stopped.');
 //            $this->moduleConfig->createLog(http_build_query($params), 'DMN params string:');
@@ -1245,7 +1250,7 @@ class Dmn extends \Magento\Framework\App\Action\Action implements \Magento\Frame
                 'DMN Callback error - there is no Order and the code did not success to create it.'
             );
             
-            http_response_code(400);
+            $this->jsonOutput->setHttpResponseCode(400);
             return 'DMN Callback error - there is no Order and the code did not success to create it.';
         }